refactor: relocate endpoint files to their own directory
