+++ src/endpoints/anime.endpoints.ts
