+++ src/endpoints/manga.endpoints.ts
