+++ src/endpoints/seasons.endpoints.ts
